Refactored CloudStore CloudStore will be in ZTS only. It will use components from common (extracted from CloudStore): - AWSCredentialsRefresher - AWSInstanceMetadataFetcher - S3ClientFactory
