+++ libs/java/server_common/src/test/java/com/yahoo/athenz/common/server/store/impl/S3ChangeLogStoreFactoryTest.java
@@ -26,8 +26,8 @@ public class S3ChangeLogStoreFactoryTest {
     @Test
     public void testCreateStore() {
         System.setProperty(ZTS_PROP_AWS_BUCKET_NAME, "s3-unit-test-bucket-name");
-        S3ChangeLogStoreFactory factory = new S3ChangeLogStoreFactory();
-        ChangeLogStore store = factory.create(null, null, null, null);
+        S3ChangeLogStoreFactory factory = new S3ChangeLogStoreFactory(null);
+        ChangeLogStore store = factory.create(null, null, null);
         assertNotNull(store);
     }
 }
